tools/vision-pipeline/py/requirements.txt
rembg
opencv-python
Pillow
numpy
tqdm
